Fix #37: better help output for negated options
Not only doesn't show the `(default: true)` for negated options, which
was quite confusing, but merges them with the line describing the
affirmative case.

Added a new test.  Since it relies on help output, it is a bit
brittle, like the other "snapshot"-based ones.  I didn't the same
execa snapshot for this because it would make a test take a lot longer
to run (becasue spawns a process).

* README.md (Options): Mention [no-]foo syntax.

* src/Command.ts (outputHelp): Consider options with negated
counterparts when printing.

* src/__test__/index.test.ts ('negated option help output'): New test.